#6 chore: 로그인 뷰컨트롤러에서 레이아웃 관련 코드는 로그인뷰로 분리
